backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
pydantic>=2.5.0
python-multipart==0.0.6
google-generativeai>=0.8.0
langchain-core>=0.3.0
tavily-python>=0.3.0
supabase>=2.0.0

